CrisisLineAI_V2/backend/requirements.txt
uvicorn==0.41.0
langchain==1.2.10
langchain-community==0.4.1
langchain-openai==1.1.10
langsmith==0.7.9
python-dotenv==1.2.1
fastapi==0.134.0
pydantic==2.11.3
langchain_ollama
assemblyai
sounddevice
